# specs/001-multitenant-foundation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Adsup Multi-tenant Foundation API
  version: 1.0.0
  description: Stable Module 1 contract. All tenant paths re-authorize active membership and scope.
servers:
  - url: http://localhost:3000
...
paths:
...
  /v1/auth/google:
    post:
      tags: [Auth]
      operationId: authenticateWithGoogle
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [idToken]
              properties:
                idToken: { type: string, minLength: 20, maxLength: 8192 }
                deviceLabel: { type: string, maxLength: 120 }
      responses:
        '200':
          description: Authenticated; profileComplete controls access to tenant mutations
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
        '422': { $ref: '#/components/responses/ValidationProblem' }
        '401': { $ref: '#/components/responses/UnauthorizedProblem' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
  /v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshSession
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [refreshToken]
              properties:
                refreshToken: { type: string, minLength: 32, maxLength: 2048 }
      responses:
        '200':
          description: Rotated token pair
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/UnauthorizedProblem' }
  /v1/auth/logout:
    post:
      tags: [Auth]
      operationId: logoutSession
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204': { description: Current session revoked }
        '401': { $ref: '#/components/responses/UnauthorizedProblem' }
  /v1/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/UnauthorizedProblem' }
    patch:
      tags: [Auth]
      operationId: confirmCurrentUserProfile
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [fullName]
              properties:
                fullName: { type: string, minLength: 2, maxLength: 120 }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '422': { $ref: '#/components/responses/ValidationProblem' }
...
  /v1/tenants/{tenantId}/form-templates:
    get:
      tags: [Forms]
      operationId: listFormTemplates
      x-permission: form.read
      parameters: [{$ref: '#/components/parameters/TenantId'}]
      responses:
        '200':
          description: Form templates
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FormTemplate' }
    post:
      tags: [Forms]
      operationId: createFormTemplate
      x-permission: form.manage
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateFormTemplate' }
      responses:
        '201':
          description: Template and initial draft created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FormTemplate' }
  /v1/tenants/{tenantId}/form-templates/{templateId}/publish:
    post:
      tags: [Forms]
      operationId: publishFormVersion
      x-permission: form.manage
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TemplateId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublishFormVersion' }
      responses:
        '201':
          description: Immutable version published
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FormVersion' }
        '422': { $ref: '#/components/responses/ValidationProblem' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
  /v1/tenants/{tenantId}/form-templates/{templateId}/archive:
    post:
      tags: [Forms]
      operationId: archiveFormTemplate
      x-permission: form.manage
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TemplateId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonCommand' }
      responses:
        '200':
          description: Template archived without deleting history
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FormTemplate' }
  /v1/tenants/{tenantId}/form-templates/{templateId}/versions/{versionId}/retire:
    post:
      tags: [Forms]
      operationId: retireFormVersion
      x-permission: form.manage
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TemplateId'
        - name: versionId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Uuid' }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReasonCommand' }
      responses:
        '200':
          description: Published version retired immutably
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FormVersion' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
  /v1/tenants/{tenantId}/form-templates/{templateId}/submissions:
    post:
      tags: [Forms]
      operationId: submitDynamicForm
      x-permission: form.submit
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TemplateId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [formVersionId, data]
              properties:
                formVersionId: { $ref: '#/components/schemas/Uuid' }
                branchId: { anyOf: [{$ref: '#/components/schemas/Uuid'}, {type: 'null'}] }
                data: { type: object }
      responses:
        '201':
          description: Validated immutable submission
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FormSubmission' }
        '422': { $ref: '#/components/responses/ValidationProblem' }
  /v1/tenants/{tenantId}/channels:
    get:
      tags: [Chat]
      operationId: listChannels
      x-permission: chat.read
      parameters: [{$ref: '#/components/parameters/TenantId'}]
      responses:
        '200':
          description: Channels current membership may access
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ChatChannel' }
    post:
      tags: [Chat]
      operationId: createChannel
      x-permission: chat.manage
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateChatChannel' }
      responses:
        '201':
          description: Channel created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatChannel' }
  /v1/tenants/{tenantId}/channels/{channelId}/messages:
    get:
      tags: [Chat]
      operationId: listMessages
      x-permission: chat.read
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Cursor-paginated messages
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessagePage' }
    post:
      tags: [Chat]
      operationId: sendMessage
      x-permission: chat.write
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [clientMessageId, body]
              properties:
                clientMessageId: { type: string, minLength: 8, maxLength: 100 }
                body: { type: string, minLength: 1, maxLength: 4000 }
                replyToMessageId: { anyOf: [{$ref: '#/components/schemas/Uuid'}, {type: 'null'}] }
      responses:
        '201':
          description: Message persisted before realtime emit
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatMessage' }
        '403': { $ref: '#/components/responses/ForbiddenProblem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
...
    TemplateId:
      name: templateId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
    ChannelId:
      name: channelId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string, maxLength: 512 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
    ValidationProblem:
      description: Invalid input
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    UnauthorizedProblem:
      description: Authentication required or invalid
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ForbiddenProblem:
      description: Actor lacks permission/scope
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    ConflictProblem:
      description: State, concurrency, last-owner or idempotency fingerprint conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
...
    User:
      type: object
      additionalProperties: false
      required: [id, fullName, profileComplete, status]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        fullName: { type: [string, 'null'] }
        profileComplete: { type: boolean }
        status: { enum: [ACTIVE, SUSPENDED, DELETED] }
    TokenPair:
      type: object
      additionalProperties: false
      required: [accessToken, accessExpiresAt, refreshToken, refreshExpiresAt]
      properties:
        accessToken: { type: string }
        accessExpiresAt: { $ref: '#/components/schemas/Timestamp' }
        refreshToken: { type: string }
        refreshExpiresAt: { $ref: '#/components/schemas/Timestamp' }
    AuthResult:
      allOf:
        - $ref: '#/components/schemas/TokenPair'
        - type: object
          required: [user]
          properties:
            user: { $ref: '#/components/schemas/User' }
...
    CreateFormTemplate:
      type: object
      additionalProperties: false
      required: [code, name]
      properties:
        code: { type: string, pattern: '^[A-Z0-9_-]{2,40}$' }
        name: { type: string, minLength: 2, maxLength: 160 }
        description: { type: string, maxLength: 1000 }
    FormTemplate:
      type: object
      required: [id, code, name, status]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        code: { type: string }
        name: { type: string }
        status: { enum: [ACTIVE, ARCHIVED] }
        currentPublishedVersionId: { anyOf: [{$ref: '#/components/schemas/Uuid'}, {type: 'null'}] }
    PublishFormVersion:
      type: object
      additionalProperties: false
      required: [jsonSchema, effectiveFrom, reason]
      properties:
        jsonSchema: { type: object }
        uiSchema: { type: [object, 'null'] }
        effectiveFrom: { $ref: '#/components/schemas/Timestamp' }
        reason: { type: string, minLength: 3, maxLength: 500 }
    ReasonCommand:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: { type: string, minLength: 3, maxLength: 500 }
    FormVersion:
      type: object
      required: [id, templateId, versionNumber, status, jsonSchema]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        templateId: { $ref: '#/components/schemas/Uuid' }
        versionNumber: { type: integer, minimum: 1 }
        status: { enum: [DRAFT, PUBLISHED, RETIRED] }
        jsonSchema: { type: object }
        effectiveFrom: { $ref: '#/components/schemas/Timestamp' }
    FormSubmission:
      type: object
      required: [id, templateId, formVersionId, submittedAt, data]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        templateId: { $ref: '#/components/schemas/Uuid' }
        formVersionId: { $ref: '#/components/schemas/Uuid' }
        submittedAt: { $ref: '#/components/schemas/Timestamp' }
        data: { type: object }
    CreateChatChannel:
      type: object
      additionalProperties: false
      required: [type, name]
      properties:
        type: { enum: [BRANCH, GROUP] }
        name: { type: string, minLength: 2, maxLength: 120 }
        branchId: { anyOf: [{$ref: '#/components/schemas/Uuid'}, {type: 'null'}] }
        membershipIds:
          type: array
          uniqueItems: true
          items: { $ref: '#/components/schemas/Uuid' }
    ChatChannel:
      type: object
      required: [id, tenantId, type, name, status]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        tenantId: { $ref: '#/components/schemas/Uuid' }
        type: { enum: [TENANT_GENERAL, BRANCH, GROUP] }
        name: { type: string }
        status: { enum: [ACTIVE, ARCHIVED] }
    ChatMessage:
      type: object
      required: [id, channelId, authorMembershipId, authorDisplayName, body, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        channelId: { $ref: '#/components/schemas/Uuid' }
        authorMembershipId: { $ref: '#/components/schemas/Uuid' }
        authorDisplayName: { type: string }
        body: { type: string }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
    MessagePage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ChatMessage' }
        nextCursor: { type: [string, 'null'] }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, correlationId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string, pattern: '^[A-Z0-9_]+$' }
        detail: { type: string }
        correlationId: { type: string }
        fieldErrors:
          type: array
          items:
            type: object
            required: [path, code, message]
            properties:
              path: { type: string }
              code: { type: string }
              message: { type: string }
security:
  - bearerAuth: []